refactor(QueryableList): move component to folder and extract highlight logic
