+++ src/pages/SendCredentials/SendCredentials.jsx
@@ -5,7 +5,7 @@ import StatusContext from '@/context/StatusContext';
 import SessionContext from '@/context/SessionContext';
 import RedirectPopup from '../../components/Popups/RedirectPopup';
 import { H1 } from '../../components/Shared/Heading';
-import QueryableList from '../../components/QueryableList';
+import QueryableList from '../../components/QueryableList/QueryableList';
 import PageDescription from '../../components/Shared/PageDescription';
 
 const Verifiers = () => {
